SIMD for gf-nf4: factorized in separated header
Extract common functions dedicated for GF-NF4 into simd_nf4.h header.
These functions use 128-bit vector instructions
