+++ src/simd_256_u32.h
@@ -297,15 +297,6 @@ mul_two_bufs(aint32* src, aint32* dest, size_t len, aint32 card = F4)
 typedef __m128i m128i;
 
 /** Return aint128 integer from a _m128i register */
-inline aint128 m128i_to_uint128(m128i v)
-{
-    aint128 i;
-    _mm_store_si128((m128i*)&i, v);
-
-    return i; // NOLINT(clang-analyzer-core.uninitialized.UndefReturn)
-}
-
-/** Return aint128 integer from a _m128i register */
 inline aint128 m256i_to_uint128(m256i v)
 {
     aint128 hi, lo;
@@ -411,104 +402,6 @@ inline void hadamard_mul(int n, aint128* _x, aint128* _y)
             _x_next_p = mul(_x_next_p, _y_p);
         }
     }
-}
-
-/* This function is the same as in simd_128_u32.h */
-inline aint128 expand16(aint16* arr, int n)
-{
-    // since n <= 4
-    uint16_t _arr[4] __attribute__((aligned(ALIGN_SIZE))) = {0, 0, 0, 0};
-    std::copy_n(arr, n, _arr);
-
-    m128i b = _mm_set_epi64(
-        _mm_setzero_si64(), _mm_set_pi16(_arr[3], _arr[2], _arr[1], _arr[0]));
-
-    return m128i_to_uint128(b);
-}
-
-/* This function is the same as in simd_128_u32.h */
-inline aint128 expand32(aint32* arr, int n)
-{
-    // since n <= 4
-    uint32_t _arr[4] __attribute__((aligned(ALIGN_SIZE))) = {0, 0, 0, 0};
-    std::copy_n(arr, n, _arr);
-
-    m128i b = _mm_set_epi32(_arr[3], _arr[2], _arr[1], _arr[0]);
-
-    return m128i_to_uint128(b);
-}
-
-/* This function is the same as in simd_128_u32.h */
-inline GroupedValues<aint128> unpack(aint128 a, int n)
-{
-    aint32 flag = 0;
-    uint32_t ai[4] __attribute__((aligned(ALIGN_SIZE)));
-    uint32_t bi[4] __attribute__((aligned(ALIGN_SIZE))) = {0, 0, 0, 0};
-    aint128 values;
-    int i;
-
-    m128i _a = _mm_loadu_si128((m128i*)&a);
-    ai[0] = _mm_extract_epi32(_a, 0);
-    ai[1] = _mm_extract_epi32(_a, 1);
-    ai[2] = _mm_extract_epi32(_a, 2);
-    ai[3] = _mm_extract_epi32(_a, 3);
-    for (i = 0; i < n; i++) {
-        if (ai[i] == 65536)
-            flag |= (1 << i);
-        else
-            bi[i] = (aint16)ai[i];
-    }
-    m128i val = _mm_set_epi64(
-        _mm_setzero_si64(), _mm_set_pi16(bi[3], bi[2], bi[1], bi[0]));
-    _mm_store_si128((m128i*)&values, val);
-
-    GroupedValues<aint128> b = {values, flag};
-
-    return b;
-}
-
-/* This function is the same as in simd_128_u32.h */
-inline aint128 pack(aint128 a)
-{
-    m128i _a = _mm_loadu_si128((m128i*)&a);
-    m128i b = _mm_set_epi32(
-        _mm_extract_epi16(_a, 3),
-        _mm_extract_epi16(_a, 2),
-        _mm_extract_epi16(_a, 1),
-        _mm_extract_epi16(_a, 0));
-
-    return m128i_to_uint128(b);
-}
-
-/* This function is the same as in simd_128_u32.h */
-inline aint128 pack(aint128 a, aint32 flag)
-{
-    aint32 b0, b1, b2, b3;
-    m128i _a = _mm_loadu_si128((m128i*)&a);
-
-    if (flag & 1)
-        b0 = 65536;
-    else
-        b0 = _mm_extract_epi16(_a, 0);
-    flag >>= 1;
-    if (flag & 1)
-        b1 = 65536;
-    else
-        b1 = _mm_extract_epi16(_a, 1);
-    flag >>= 1;
-    if (flag & 1)
-        b2 = 65536;
-    else
-        b2 = _mm_extract_epi16(_a, 2);
-    flag >>= 1;
-    if (flag & 1)
-        b3 = 65536;
-    else
-        b3 = _mm_extract_epi16(_a, 3);
-
-    m128i b = _mm_set_epi32(b3, b2, b1, b0);
-
-    return m128i_to_uint128(b);
 }
 
 } // namespace u32
